Extract superclass BaseGradingSystemComponent from Interval and Detailed components.
