+++ src/main/webapp/app/grading-system/base-grading-system/_base-grading-system.component.scss
@@ -1,0 +1,18 @@
+.dropdown-container {
+    display: flex;
+    justify-content: space-between;
+    width: 100%;
+    margin: 1em 0;
+}
+
+.dropdown-container .form-group .form-control {
+    margin-top: 0.25em;
+}
+
+.grading-scale-table-row td {
+    width: 17%;
+
+    input {
+        min-width: 50%;
+    }
+}
