+++ src/main/webapp/app/grading-system/detailed-grading-system/detailed-grading-system.component.scss
@@ -1,23 +1,6 @@
-.dropdown-container {
-    display: flex;
-    justify-content: space-between;
-    width: 100%;
-    margin: 1em 0;
-}
+@import 'src/main/webapp/app/grading-system/base-grading-system/base-grading-system.component';
 
 .dropdown-container .form-group {
     width: 23%;
     min-width: 20%;
-
-    .form-control {
-        margin-top: 0.25em;
-    }
-}
-
-.grading-scale-table-row td {
-    width: 17%;
-
-    input {
-        min-width: 50%;
-    }
 }
